fix: address bug with Link to GitHub not triggering
Because the watch function only triggers when the observed value changes after the component has been mounted, watch was not picking up on the GitHub access token query parameter, because it technically was not *changed* -- it was already present when the component was mounted after the GitHub redirect.

Adding `onMounted` ensures that the logic triggers on page load. Because watch was not working and is not needed, it has been removed.

Additionally, after loading under these conditions, the gh_access_token query parameter is now removed, just to keep things clean 🧹